+++ tests/python/test_async.py
@@ -34,3 +34,21 @@ def test_numpy():
 
     for i in range(n):
         assert x[i] == i * 10
+
+
+@ti.test(require=ti.extension.async_mode, async_mode=True)
+def test_listgen_opt_with_offsets():
+    x = ti.field(dtype=ti.i32)
+
+    ti.root.pointer(ti.i, 4).dense(ti.i, 4).place(x, offset=-8)
+
+    @ti.kernel
+    def inc():
+        for i in x:
+            x[i] += 1
+
+    for i in range(10):
+        inc()
+
+    ti.sync()
+    assert ti.get_kernel_stats().get_counters()['launched_tasks_list_gen'] <= 2
